feat(cli): RewriteManifests maintenance task
Rebased onto upstream/main. Upstream refactored the shared commit() path
(buildManifests/addedContent, summaryOnRetry, accumulateSummaryDelta), which
overlapped the branch's own commit() changes, so the 6 review-iteration commits
were replayed as one reconciled diff rather than conflicting per-commit.

Adopts upstream's addedContent model: it subsumes the branch's computeOwnManifests
/ inheritedOnlyRebuilder scaffolding (a rewrite has empty addedContent, so OCC
retries re-derive from the fresh parent for free). Keeps the branch's
commitManifests split, supersededAccumulator cleanup, and rewrite-only
partition-summary-limit override.

+++ table/snapshot_producers_test.go
@@ -857,7 +857,7 @@ func TestManifestMergeGroupLimitsConcurrentBins(t *testing.T) {
 	}
 
 	mgr := manifestMergeManager{
-		targetSizeBytes:  int(manifests[0].Length() * 2),
+		targetSizeBytes:  manifests[0].Length() * 2,
 		minCountToMerge:  1,
 		mergeEnabled:     true,
 		mergeConcurrency: mergeConcurrency,
